Diagnostics Fix

+++ Gamer.Diagnostics/Main.cs
@@ -176,6 +176,8 @@ namespace Gamer.Diagnostics
 
         internal static void AnalizeContent(string file)
         {
+            if (!File.Exists(file))
+                return;
             var result = AnalizeContent(File.ReadAllLines(file), DateTime.Now.AddHours(-1));
             File.WriteAllText(Path.Combine(Path.GetDirectoryName(file), Path.GetFileNameWithoutExtension(file)+ ".analized.log"), Newtonsoft.Json.JsonConvert.SerializeObject(result));
             File.Delete(file);
